+ bugfix imports everywhere + moved fileio and snakeylowercaser to utils to resolve circular imports
